Refactored Python View to remove dependence on matplotlib
In principal, image data to be displayed in a view render window could
be generated many different ways in Python - using matplotlib, OpenCV,
or even direct manipulation of a numpy array. In light of that, this
change decouples the Python View from matplotlib. Matplotlib may still
be used to generate images in the Python view with slight modification
to existing Python scripts used in the Python view.

From a user's perspective, the only changes to know about are how to
define the "render" function in the Python script. Previously, the
"render" function had the signature

  def render(view, figure)

where "view" is the Python View instance calling this function and
"figure" was a matplotlib figure that was set up by the view
automatically. The function was not expected to return a value.

Now the "render" function has the signature

  def render(view, width, height)

where "view" is still the Python View and the "width" and "height"
parameters are the width and height of the current view render
window. IMPORTANT - the function is now expected to return a
vtkImageData object containing the image content to display in the
view render window.

To generate an image from matplotlib, Figure and FigureCanvas
object need to be created the following way:

  from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
  from matplotlib.figure import Figure

  figure = Figure()
  figureCanvas = FigureCanvas(figure)
  figure.set_dpi(72)
  figure.set_size_inches(float(width)/72.0, float(height)/72.0)

This "figure" object can be used in the same way as the "figure"
argument in the previous signature of the "render" method.

A convenience function has been added to to the python_view module to
create a vtkImageData object from a matplotlib Figure object. It is
used as follows:

  from paraview import python_view
  vtk_image = python_view.figure_to_image(figure)

This refactoring has simplified the implementation of the Python View
dramatically. The matplotlib utilities are no longer needed with the
changes in the Python View. Furthermore, Python View is now available
whenever Python is enabled.

Change-Id: Ic41301f5f7492fa8f171570b03eff317c7faf108